Split Cloning System into Utility
